Update README.md
The preferred author identifier is the [ORCID](https://orcid.org/) identifier. If the author does not possess an ORCID identifier, an INSPIRE-ID may be used as an alternative. A persistent ID for an author allows the INSPIRE service team to identify authors and attach the respective identifiers to the author ID system.